+++ src/main/javascript/Semantic/index.js
@@ -1,2 +1,3 @@
 export { default as Block } from './Block';
 export { default as Section } from './Section';
+export { default as Button } from './Button';